Enable integration tests against OpenSearch
This PR removes all tests and client features that relied on xpack features.

It adds a workflow that runs integration tests on push and pull requests.

The tests run against all versions of OpenSearch that have the required plugins.

Integration tests use the following OpenSearch plugins:
- ingest-attachment,
- analysis-kuromoji,
- analysis-icu,
- analysis-phonetic,
- mapper-murmur3,
- analysis-nori

OpenSearch 1.0.0 and 1.0.1 are excluded because plugin artifacts are not available for download.

Signed-off-by: MaxKsyunz <maxk@bitquilltech.com>